+++ src/primitive/mod.rs
@@ -158,6 +158,7 @@ impl fmt::Display for ImplPrimitive {
         use ImplPrimitive::*;
         use Primitive::*;
         match self {
+            UnPop => write!(f, "{Un}{Pop}"),
             UnBits => write!(f, "{Un}{Bits}"),
             UnWhere => write!(f, "{Un}{Where}"),
             UnCouple => write!(f, "{Un}{Couple}"),
@@ -369,7 +370,6 @@ impl Primitive {
             Coordinate
                 | (This | Recur)
                 | (Rectify | All | Cascade | By)
-                | (Map | Insert | Has | Get | Remove)
                 | Bind
                 | (Shapes | Types)
                 | Sys(SysOp::FFI)
@@ -845,6 +845,13 @@ impl Primitive {
 impl ImplPrimitive {
     pub(crate) fn run(&self, env: &mut Uiua) -> UiuaResult {
         match self {
+            ImplPrimitive::UnPop => {
+                env.push(
+                    env.value_fill()
+                        .ok_or_else(|| env.error("No fill set").fill())?
+                        .clone(),
+                );
+            }
             ImplPrimitive::Asin => env.monadic_env(Value::asin)?,
             ImplPrimitive::UndoKeep => {
                 let from = env.pop(1)?;